bin/faucet-tool: introduce .hnsFaucetTool
The tool creates a back up the of the wallet details
printed to the screen. It also produces a QR code which
encodes the encrypted address. These items will be stored
under the user's home directory in a dir named '.hnsFaucet'.
The filenames will be prefixed with the generated
receiving address.